requirements.txt
fastapi
uvicorn
pymongo
cohere
qdrant-client
pydantic-settings
PyMuPDF
python-dotenv
python-multipart
